relocate autosave functionality to actionTypesStore
